+++ app/controllers/account_controller.rb
@@ -9,10 +9,10 @@ class AccountController < ApplicationController
   def dashboard
     @projects = current_user.all_viewable_and_site_projects
                             .by_favorite(current_user.id).unarchived
-                            .order("(favorited IS NULL or favorited = 'f') ASC, position, name")
+                            .order(Arel.sql("(favorited IS NULL or favorited = 'f') ASC, position, name"))
                             .page(params[:page]).per(Project::PER_PAGE)
     @favorited_projects = @projects.where(project_preferences: { favorited: true })
-    @current_projects = @projects.where(project_preferences: { favorited: [false, nil] }).reorder("lower(name) asc")
+    @current_projects = @projects.where(project_preferences: { favorited: [false, nil] }).reorder(Arel.sql("lower(name) asc"))
     redirect_to @projects.first if current_user.all_viewable_and_site_projects.count == 1
   end