+++ app/controllers/api/v1/surveys_controller.rb
@@ -10,6 +10,8 @@ class Api::V1::SurveysController < Api::V1::BaseController
   # GET /api/v1/projects/1-AUTHENTICATION_TOKEN/subjects/1/surveys/:event/:design.json
   def show_survey
     if @event && @design && @subject_event && @sheet
+      # @sheet.check_response_count_change
+      # @sheet.update_associated_subject_events!
       render :survey
     else
       head :no_content
@@ -48,15 +50,8 @@ class Api::V1::SurveysController < Api::V1::BaseController
         true
       end
     if save_result
-      # @sheet.audit_set_lasted_edited!
-      # @sheet.update_response_count!
-      # @sheet.update_associated_subject_events!
-      # @sheet.subject_event.update_coverage! if @sheet.subject_event
-
       @sheet.audit_set_lasted_edited!
       @sheet.reset_response_count!
-      @sheet.update_associated_subject_events!
-
       render json: {}, status: :ok
     else
       render :survey_page, status: :unprocessable_entity